Jogo Joguifero
Começo da classe Personagem e testes com o Gerenciador Grafico

+++ Megaman++/Megaman++/Entidades/Entidade.cpp
@@ -4,7 +4,7 @@ Entidade::Entidade(): x(-1), y(-1)
 {
 }
 
-Entidade::Entidade(int xI, int yI): x(xI), y(yI)
+Entidade::Entidade(float xI, float yI): x(xI), y(yI)
 {
 }
